+++ lib/flumpose.dart
@@ -18,6 +18,7 @@ export 'src/flex_extensions.dart';
 export 'src/stack_extensions.dart';
 export 'src/responsive_extensions.dart';
 export 'src/utility_extensions.dart';
+export 'src/form_extensions.dart';
 
 // Performance optimizations
 export 'src/decoration_builder.dart';